-updated profile fragement
updated profile fragment and deleted the activity that was there before.
