+++ app/src/main/AndroidManifest.xml
@@ -12,10 +12,6 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Vault"
         tools:targetApi="31">
-        <activity
-            android:name=".MainActivity"
-            android:exported="true" />
-
         <meta-data
             android:name="com.google.android.gms.version"
             android:value="@integer/google_play_services_version" />
@@ -24,7 +20,7 @@
             android:value="@string/API_KEY" />
 
         <activity
-            android:name=".ui.profile.ProfileManagement"
+            android:name=".MainActivity"
             android:exported="true"
             >
             <intent-filter>
