+++ app/src/main/java/com/example/vault/ui/profile/ProfileFragment.kt
@@ -1,5 +1,4 @@
 package com.example.vault.ui.profile
-
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -7,9 +6,16 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
+import com.example.vault.R
 import com.example.vault.databinding.FragmentProfileBinding
+import com.google.android.gms.maps.CameraUpdateFactory
+import com.google.android.gms.maps.GoogleMap
+import com.google.android.gms.maps.OnMapReadyCallback
+import com.google.android.gms.maps.SupportMapFragment
+import com.google.android.gms.maps.model.LatLng
+import com.google.android.gms.maps.model.MarkerOptions
 
-class ProfileFragment : Fragment() {
+class ProfileFragment : Fragment(), OnMapReadyCallback {
 
     private var _binding: FragmentProfileBinding? = null
 
@@ -23,12 +29,20 @@ class ProfileFragment : Fragment() {
 
         _binding = FragmentProfileBinding.inflate(inflater, container, false)
         val root: View = binding.root
-
-//        val textView: TextView = binding.textNotifications
-//        profileViewModel.text.observe(viewLifecycleOwner) {
-//            textView.text = it
-//        }
+        val mapFragment = childFragmentManager.findFragmentById(R.id.map) as? SupportMapFragment
+        mapFragment?.getMapAsync(this)
         return root
+    }
+
+    override fun onMapReady(map: GoogleMap) {
+
+        val university = LatLng(43.9456, -78.8968)
+        map.addMarker(
+            MarkerOptions()
+                .position(university)
+                .title("Ontario Tech University")
+        )
+        map.moveCamera(CameraUpdateFactory.newLatLngZoom(university, 15f))
     }
 
     override fun onDestroyView() {
