--- app/src/main/res/layout/activity_profile_management.xml
@@ -1,106 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/main_content"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <LinearLayout
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            app:layout_constraintWidth_percent="0.95"
-            android:orientation="vertical"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent">
-            <TextView
-                android:id="@+id/app_title"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/app_name"
-                android:textSize="@dimen/header_1"
-                android:textStyle="bold"/>
-
-            <TextView
-                android:id="@+id/setup_license"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/setup"
-                android:textSize="@dimen/header_2"
-                android:textStyle="bold"
-                android:gravity="center"
-                android:layout_marginBottom="@dimen/activity_vertical_margin"/>
-
-            <TextView
-                android:id="@+id/login_signup"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/login_signup"
-                android:textSize="@dimen/header_3"
-                android:layout_marginBottom="@dimen/activity_vertical_margin"/>
-
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:orientation="vertical"
-                android:gravity="center_horizontal"
-                tools:ignore="UseCompoundDrawables"
-                android:layout_marginBottom="@dimen/activity_vertical_margin">
-
-                <ImageView
-                    android:id="@+id/github_icon"
-                    android:layout_width="match_parent"
-                    android:layout_height="45dp"
-                    android:contentDescription="@string/github"
-                    app:srcCompat="@drawable/github" />
-
-                <Button
-                    android:id="@+id/github_text"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:drawablePadding="20dp"
-                    android:gravity="center"
-                    android:text="@string/github"
-                    android:textStyle="bold"
-                    android:padding="8dp"
-                    android:backgroundTint="@color/github_green"
-                    android:textColor="@android:color/white"
-                    android:letterSpacing="0.1"/>
-            </LinearLayout>
-
-            <TextView
-                android:id="@+id/contact_us"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/contact_us"
-                android:layout_marginBottom="@dimen/activity_vertical_margin"
-                android:textSize="@dimen/header_2"
-                android:textStyle="bold"/>
-
-            <TextView
-                android:id="@+id/school_name"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/school_name"
-                android:textSize="@dimen/header_3"/>
-
-            <androidx.fragment.app.FragmentContainerView
-                android:id="@+id/map"
-                android:name="com.google.android.gms.maps.SupportMapFragment"
-                android:layout_width="match_parent"
-                android:layout_height="300dp"
-                tools:context="com.example.mapwithmarker.MapsMarkerActivity" />
-        </LinearLayout>
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
